+++ era_seo_manager/controllers/__init__.py
@@ -1,1 +1,1 @@
-from . import main, sitemap, robots, redirect, blog, feed
+from . import main, sitemap, robots, redirect