+++ era_seo_manager/models/blog_category.py
@@ -1,9 +1,0 @@
-from odoo import fields, models
-
-
-class EraBlogCategory(models.Model):
-    _name = 'era.blog.category'
-    _description = 'Blog Category'
-
-    name = fields.Char(string='Name', required=True, translate=True)
-    active = fields.Boolean(default=True)