+++ era_seo_manager/models/blog_series.py
@@ -1,9 +1,0 @@
-from odoo import fields, models
-
-
-class EraBlogSeries(models.Model):
-    _name = 'era.blog.series'
-    _description = 'Blog Series'
-
-    name = fields.Char(string='Name', required=True, translate=True)
-    active = fields.Boolean(default=True)